requirements.txt
# --- GUI Framework ---
PySide6>=6.10.1,<7  # Qt6-only, avoid Qt7 breaking changes

# --- Audio Processing ---
pyannote.audio>=4.0.3
soundfile>=0.13.1
langdetect>=1.0.9
scipy>=1.16.3
numpy>=2.0.0

# --- AI & Machine Learning ---
transformers>=4.57.3
sentence-transformers>=5.1.2

# --- LLM Clients ---
openai>=2.8.1
google-genai>=1.53.0

# --- Utilities ---
tqdm>=4.67.1
onnx>=1.17.0
noisereduce